+++ src/net.rs
@@ -14,21 +14,25 @@
 
 use crate::{Error, Result, ReturnCode};
 use phidget_sys as ffi;
-use std::{ffi::CString, os::raw::c_int};
+use std::{
+    ffi::{CStr, CString},
+    os::raw::{c_char, c_int, c_void},
+};
 
 /// Phidget server types
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
 #[repr(u32)]
 #[allow(missing_docs)]
 pub enum ServerType {
+    #[default]
     None = ffi::PhidgetServerType_PHIDGETSERVER_NONE, // 0
     DeviceListener = ffi::PhidgetServerType_PHIDGETSERVER_DEVICELISTENER, // 1
-    Device = ffi::PhidgetServerType_PHIDGETSERVER_DEVICE, // 2
-    DeviceRemote = ffi::PhidgetServerType_PHIDGETSERVER_DEVICEREMOTE, // 3
-    WwwListener = ffi::PhidgetServerType_PHIDGETSERVER_WWWLISTENER, // 4
-    Www = ffi::PhidgetServerType_PHIDGETSERVER_WWW,   // 5
-    WwwRemote = ffi::PhidgetServerType_PHIDGETSERVER_WWWREMOTE, // 6
-    Sbc = ffi::PhidgetServerType_PHIDGETSERVER_SBC,   // 7
+    Device = ffi::PhidgetServerType_PHIDGETSERVER_DEVICE,                 // 2
+    DeviceRemote = ffi::PhidgetServerType_PHIDGETSERVER_DEVICEREMOTE,     // 3
+    WwwListener = ffi::PhidgetServerType_PHIDGETSERVER_WWWLISTENER,       // 4
+    Www = ffi::PhidgetServerType_PHIDGETSERVER_WWW,                       // 5
+    WwwRemote = ffi::PhidgetServerType_PHIDGETSERVER_WWWREMOTE,           // 6
+    Sbc = ffi::PhidgetServerType_PHIDGETSERVER_SBC,                       // 7
 }
 
 impl TryFrom<u32> for ServerType {
@@ -47,6 +51,57 @@ impl TryFrom<u32> for ServerType {
             ffi::PhidgetServerType_PHIDGETSERVER_SBC => Ok(Sbc),   // 7
             _ => Err(ReturnCode::InvalidArg),
         }
+    }
+}
+
+// Converts the C char ptr to a string, returning an empty string on
+// a null ptr, or a UTF-8 error on conversion failure.
+fn ptr_to_string(p: *const c_char) -> std::result::Result<String, std::str::Utf8Error> {
+    let s = unsafe {
+        match p.as_ref() {
+            Some(s) => CStr::from_ptr(s).to_str()?.to_owned(),
+            None => String::new(),
+        }
+    };
+    Ok(s)
+}
+
+/// Information about a phidget server
+#[derive(Default, Debug)]
+pub struct Server {
+    /// The server name
+    pub name: String,
+    /// The server type
+    pub typ: ServerType,
+    /// Flags
+    pub flags: u32,
+    /// The network address
+    pub addr: String,
+    /// The host name
+    pub host: String,
+    /// The port
+    pub port: u16,
+}
+
+impl TryFrom<&ffi::PhidgetServer> for Server {
+    type Error = Error;
+
+    fn try_from(srvr: &ffi::PhidgetServer) -> Result<Self> {
+        let name = ptr_to_string(srvr.name)?;
+        let typ = ServerType::try_from(srvr.type_)?;
+        let flags = srvr.flags as u32;
+        let addr = ptr_to_string(srvr.addr)?;
+        let host = ptr_to_string(srvr.host)?;
+        let port = srvr.port as u16;
+
+        Ok(Self {
+            name,
+            typ,
+            flags,
+            addr,
+            host,
+            port,
+        })
     }
 }
 
@@ -116,38 +171,90 @@ pub fn set_server_passward(server_name: &str, password: &str) -> Result<()> {
 /// the network.
 /// Currently Multicast DNS is used to discover and publish Phidget servers.
 pub fn enable_server_discovery(server_type: ServerType) -> Result<()> {
-    ReturnCode::result(unsafe { ffi::PhidgetNet_enableServerDiscovery(server_type as u32) })
+    ReturnCode::result(unsafe {
+        ffi::PhidgetNet_enableServerDiscovery(server_type as ffi::PhidgetServerType)
+    })
 }
 
 /// Disables the dynamic discovery of servers that publish their identity.
 /// This does not disconnect already established connections.
 pub fn disable_server_discovery(server_type: ServerType) -> Result<()> {
-    ReturnCode::result(unsafe { ffi::PhidgetNet_disableServerDiscovery(server_type as u32) })
+    ReturnCode::result(unsafe {
+        ffi::PhidgetNet_disableServerDiscovery(server_type as ffi::PhidgetServerType)
+    })
+}
+
+/// Callback when a server is added
+pub type ServerAddedCallback = dyn Fn(Server) + Send + 'static;
+
+/// Callback when a server is removed
+pub type ServerRemovedCallback = dyn Fn(Server) + Send + 'static;
+
+// Low-level, unsafe, callback for when a server is added
+// The context is a double-boxed pointer to the safe Rust callback.
+unsafe extern "C" fn on_server_added(
+    ctx: *mut c_void,
+    srvr: ffi::PhidgetServerHandle,
+    // TODO: What is this?
+    _kv: *mut c_void,
+) {
+    println!("on_server_added");
+    if ctx.is_null() {
+        return;
+    }
+
+    let cb: &mut Box<ServerAddedCallback> = &mut *(ctx as *mut _);
+    let srvr = srvr
+        .as_ref()
+        .and_then(|s| Server::try_from(s).ok())
+        .unwrap_or_default();
+    cb(srvr);
+}
+
+/// Assigns a handler to be called when a "server added" event occurs.
+pub fn set_on_server_added_handler<F>(cb: F) -> Result<()>
+where
+    F: Fn(Server) + Send + 'static,
+{
+    // 1st box is fat ptr, 2nd is regular pointer.
+    let cb: Box<Box<ServerAddedCallback>> = Box::new(Box::new(cb));
+    let ctx = Box::into_raw(cb) as *mut c_void;
+
+    ReturnCode::result(unsafe {
+        ffi::PhidgetNet_setOnServerAddedHandler(Some(on_server_added), ctx)
+    })
+}
+
+// Low-level, unsafe, callback for when a server is removed
+// The context is a double-boxed pointer to the safe Rust callback.
+unsafe extern "C" fn on_server_removed(ctx: *mut c_void, srvr: ffi::PhidgetServerHandle) {
+    if ctx.is_null() {
+        return;
+    }
+
+    let cb: &mut Box<ServerRemovedCallback> = &mut *(ctx as *mut _);
+    let srvr = srvr
+        .as_ref()
+        .and_then(|s| Server::try_from(s).ok())
+        .unwrap_or_default();
+    cb(srvr);
+}
+
+/// Assigns a handler to be called when a "server removed" event occurs.
+pub fn set_on_server_removed_handler<F>(cb: F) -> Result<()>
+where
+    F: Fn(Server) + Send + 'static,
+{
+    // 1st box is fat ptr, 2nd is regular pointer.
+    let cb: Box<Box<ServerRemovedCallback>> = Box::new(Box::new(cb));
+    let ctx = Box::into_raw(cb) as *mut c_void;
+
+    ReturnCode::result(unsafe {
+        ffi::PhidgetNet_setOnServerRemovedHandler(Some(on_server_removed), ctx)
+    })
 }
 
 /*
-pub type PhidgetNet_OnServerAddedCallback = ::std::option::Option<
-    unsafe extern "C" fn(
-        ctx: *mut ::std::os::raw::c_void,
-        server: PhidgetServerHandle,
-        kv: *mut ::std::os::raw::c_void,
-    ),
->;
-pub type PhidgetNet_OnServerRemovedCallback = ::std::option::Option<
-    unsafe extern "C" fn(ctx: *mut ::std::os::raw::c_void, server: PhidgetServerHandle),
->;
-extern "C" {
-    pub fn PhidgetNet_setOnServerAddedHandler(
-        fptr: PhidgetNet_OnServerAddedCallback,
-        ctx: *mut ::std::os::raw::c_void,
-    ) -> PhidgetReturnCode;
-}
-extern "C" {
-    pub fn PhidgetNet_setOnServerRemovedHandler(
-        fptr: PhidgetNet_OnServerRemovedCallback,
-        ctx: *mut ::std::os::raw::c_void,
-    ) -> PhidgetReturnCode;
-}
 extern "C" {
     pub fn PhidgetNet_getServerAddressList(
         hostname: *const ::std::os::raw::c_char,